04_langchain_trusted_rag/requirements.txt
dokis[langchain]
langchain
langchain-openai
langchain-community
python-dotenv
pydantic
rich